Code elegance.
